- Changed vector and matrix classes "NUM" field from 'u32' to 'size_t' for portability. - Added some documentation to quaternion class. - Abstracted "quaternion::set/getMatrix" functions (to be used for matrix4 and matrix3 classes).

+++ trunk/sources/Base/spDimensionVector3D.hpp
@@ -36,7 +36,7 @@ template <typename T> class vector3d
     public:
         
         //! Number of components (3). This is used for vector arithmetic templates.
-        static const u32 NUM = 3;
+        static const size_t NUM = 3;
         
         vector3d() :
             X(0),
@@ -163,11 +163,11 @@ template <typename T> class vector3d
         
         /* === Additional operators === */
         
-        inline const T& operator [] (u32 i) const
+        inline const T& operator [] (size_t i) const
         {
             return *(&X + i);
         }
-        inline T& operator [] (u32 i)
+        inline T& operator [] (size_t i)
         {
             return *(&X + i);
         }
